backend/requirements.txt
fastapi
uvicorn
supabase
python-multipart
redis
celery
openai
pydantic-settings
langchain-openai
langchain-text-splitters
langchain-community
tiktoken
unstructured[pdf]
python-magic
python-multipart
pyjwt
beautifulsoup4
requests
